+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for symbolizing addresses in a vDSO
+  - Added `vdso` attribute to `symbolize::Process` type
 - Adjusted debug link resolution to handle self-referential debug links
   more gracefully
 - Report "symbolic" path in `symbolize::Sym::module` when using process